backend/requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
python-multipart==0.0.5
pillow==8.3.1
numpy==1.21.2
fuzzywuzzy==0.18.0
python-Levenshtein==0.12.2
azure-cognitiveservices-vision-computervision==0.9.0
msrest==0.7.1
requests==2.26.0
